+++ src/transformer/index.ts
@@ -30,8 +30,8 @@ import { metadataDecorator } from './metadata-decorator';
 import { rtHelper } from './rt-helper';
 import { serialize } from './serialize';
 import * as ts from 'typescript';
+import { T_ANY, T_ARRAY, T_INTERSECTION, T_THIS, T_TUPLE, T_UNION, T_UNKNOWN, T_GENERIC } from '../common';
 import { cloneEntityNameAsExpr, dottedNameToExpr, entityNameToString, getRootNameOfEntityName } from './utils';
-import { T_ANY, T_ARRAY, T_INTERSECTION, T_THIS, T_TUPLE, T_UNION, T_UNKNOWN } from '../common';
 
 export enum TypeReferenceSerializationKind {
     // The TypeReferenceNode could not be resolved.
@@ -162,6 +162,28 @@ const transformer: (program : ts.Program) => ts.TransformerFactory<ts.SourceFile
                 if (!typeNode)
                     return ts.factory.createVoidZero();
                 
+                let expr = serializeBaseTypeRef(typeNode, extended);
+
+                if (extended) {
+                    if (ts.isTypeReferenceNode(typeNode)) {
+                        if (typeNode.typeArguments && typeNode.typeArguments.length > 0) {
+                            // Handle generic types like Promise<string> etc
+                            expr = serialize({ 
+                                TΦ: T_GENERIC, 
+                                t: literalNode(expr),
+                                p: typeNode.typeArguments.map(x => literalNode(serializeTypeRef(x, extended)))
+                            });
+                        }
+                    }
+                }
+
+                return expr;
+            }
+
+            function serializeBaseTypeRef(typeNode : ts.Node, extended): ts.Expression {
+                if (!typeNode)
+                    return ts.factory.createVoidZero();
+                
 
                 if (ts.isTypeReferenceNode(typeNode)) {
                     const resolver = context['getEmitResolver']();
@@ -170,7 +192,7 @@ const transformer: (program : ts.Program) => ts.TransformerFactory<ts.SourceFile
                     let expr : ts.PropertyAccessExpression | ts.Identifier ;
 
                     if (ts.isIdentifier(typeNode.typeName)) {
-                        let primitiveTypes = ['Number', 'String', 'Boolean', 'Function', 'Object'];
+                        let primitiveTypes = ['Number', 'String', 'Boolean', 'Function', 'Object', 'Promise', 'Symbol'];
                         if (primitiveTypes.includes(typeNode.typeName.text)) {
                             return ts.factory.createIdentifier(typeNode.typeName.text);
                         }
@@ -254,6 +276,24 @@ const transformer: (program : ts.Program) => ts.TransformerFactory<ts.SourceFile
                         return ts.factory.createIdentifier('Array');
                 }
                 
+                if (ts.isLiteralTypeNode(typeNode)) {
+                    let literal = typeNode.literal;
+
+                    if (ts.isLiteralExpression(literal))
+                        return literal;
+                    if (ts.isPrefixUnaryExpression(literal))
+                        return literal;
+                    
+                    switch (literal.kind) {
+                        case ts.SyntaxKind.NullKeyword:
+                            return ts.factory.createIdentifier('null');
+                        case ts.SyntaxKind.FalseKeyword:
+                            return ts.factory.createIdentifier('false');
+                        case ts.SyntaxKind.TrueKeyword:
+                            return ts.factory.createIdentifier('true');
+                    }
+                }
+                
                 if (ts.isTupleTypeNode(typeNode)) {
                     if (!extended)
                         return ts.factory.createIdentifier('Object');
@@ -289,17 +329,17 @@ const transformer: (program : ts.Program) => ts.TransformerFactory<ts.SourceFile
                     });
                 }
 
-                if (ts.isThisTypeNode(typeNode)) {
+                if (ts.isThisTypeNode(typeNode))
                     return serialize({ TΦ: T_THIS });
-                }
 
-                if (ts.isConditionalTypeNode(typeNode)) {
+                if (ts.isConditionalTypeNode(typeNode))
                     return ts.factory.createIdentifier('Object');
-                }
 
-                if (ts.isTypePredicateNode) {
+                if (ts.isTypePredicateNode(typeNode))
                     return ts.factory.createIdentifier('Boolean');
-                }
+
+                if (typeNode.kind === ts.SyntaxKind.UndefinedKeyword)
+                    return ts.factory.createVoidZero();
 
                 /// ??