+++ CHANGELOG.md
@@ -1,8 +1,29 @@
+# v0.0.19
+
+- Added better handling for literal types to `ReflectedClass`.
+    * You can now expect `isClass(Boolean)` to be true for types `true` and `false`, `isClass(Object)` to be true for 
+      `null`, `isClass(Number)` to be true for numeric literals and `isClass(String)` to be true for string literals.
+    * Added `isLiteral(value)` to check for a literal value
+- Fixed a bug where all unknown types were reported as `Boolean`
+- Added support for `undefined` type
+- Added a number of helpers for checking for literal types to `ReflectedTypeRef`
+
+# v0.0.18
+
+- Added support for type literal types, ie `foo(bar : false, baz : null, foobar : 123)`
+
+# v0.0.17
+
+- Fix: do not crash when property has no type (https://github.com/rezonant/typescript-rtti/commit/474eddf15160457e57a786f0c67918e99a11d8c2)
+
 # v0.0.15
+
+**Features**
+- Added support for serializing generic types including their type arguments. This means you can now obtain the type of a `Promise` for instance (provided that the referenced type has a value at runtime). Additionally, cases where the type references an interface, and that interface has type parameters will now emit a generic type which exposes the types of the parameters, even if the interface itself does not have a runtime value. For instance `InterfaceA<InterfaceB>` would emit a generic type with base type `Object` and one parameter type of `Object`.
 
 **Breaking**
 - Made the structure of the `RtTypeRef` family of interfaces internal along with creation of `ReflectedTypeRef` and its `ref` property.
-  Technically this is a breaking change, but these interfaces have only been exposed since v0.0.14, which was released earlier today.
+  Technically this is a breaking change, but these interfaces have only been exposed since v0.0.14
 
 # v0.0.14